Fixed assoications

+++ server/db/models/product.js
@@ -35,11 +35,11 @@ const Product = db.define('product', {
     type: Sequelize.TEXT
   },
   price: {
-    type: Sequelize.DECIMAL,
+    type: Sequelize.INTEGER,
     allowNull: false
   },
   rating: {
-    type: Sequelize.DECIMAL
+    type: Sequelize.INTEGER
   },
   quantity: {
     type: Sequelize.INTEGER,
